nasa-asteroids-tracker-final/requirements.txt
streamlit
pandas
mysql-connector-python
requests